requirements.txt
python-telegram-bot
python-bitcoinrpc
asyncpg
